### 2025-04-06 #8008 @editorialbot
[PRE REVIEW]: Vivid Volcano: Empowering Non-Bioinformaticians to Analyze Pre-Processed Omics   Data
**Submitting author:** @DatViseR (<a href="http://orcid.org/0000-0002-8454-4421">Tomasz Stępkowski</a>)
**Repository:** https://github.com/DatViseR/Vivid-Volcano
**Branch with paper.md** (empty if default branch): master
**Version:** Version: 0.1.0
