### 2025-04-10 4c06c6c
Improved the speed of upload by introducing the data.table::fread() insted of read_delim(), increased the max file size to 40mb
