### 2025-04-10 b7e4766
Initial commit with telemetry branch for deployment with uplouding files speeded up by using fread
