### 2025-04-10 050888f
Change the R scripts catalog name to R to be coherent with posit cloud environment
