### 2025-04-10 d869ed7
changed the error in file name with vivid_functions.R
