### 2025-04-16 97039bf
Corrected the bib file to contain all missing dois
