+++ Publication/paper.bib
@@ -14,82 +14,94 @@
 	langid = {en}
 }
 
-@article{shiny,
-	title = {shiny: Web Application Framework for R},
-	author = {Chang, Winston and Cheng, Joe and Allaire, JJ and Sievert, Carson and Schloerke, Barret and Xie, Yihui and Allen, Jeff and McPherson, Jonathan and Dipert, Alan and Borges, Barbara},
-	year = {2024},
-	date = {2024},
-	url = {https://CRAN.R-project.org/package=shiny}
+@Manual{shiny,
+  title = {shiny: Web Application Framework for R},
+  author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+  year = {2024},
+  note = {R package version 1.8.1},
+  url = {https://CRAN.R-project.org/package=shiny},
+  doi = {10.32614/CRAN.package.shiny},
 }
 
-@article{shiny.semantic,
-	title = {shiny.semantic: Semantic UI Support for Shiny},
-	author = {Stachura, Filip and Krzeminski, Dominik and Igras, Krystian and Forys, Adam and {Przytu{\l}a}, {Pawe{\l}} and Chojna, Jakub and Mierzwa-Sulima, Olga and Nowicki, Jakub and Makowski, Tymoteusz},
-	year = {2024},
-	date = {2024},
-	url = {https://appsilon.github.io/shiny.semantic/}
+@Manual{shiny.semantic,
+  title = {shiny.semantic: Semantic UI Support for Shiny},
+  author = {Filip Stachura and Dominik Krzeminski and Krystian Igras and Adam Forys and {Paweł Przytuła} and Jakub Chojna and Olga Mierzwa-Sulima and Jakub Nowicki and Tymoteusz Makowski},
+  year = {2024},
+  note = {R package version 0.4.4},
+  url = {https://CRAN.R-project.org/package=shiny.semantic},
+  doi = {10.32614/CRAN.package.shiny.semantic},
 }
 
-@article{DT,
-	title = {DT: A Wrapper of the JavaScript Library 'DataTables'},
-	author = {Xie, Yihui and Cheng, Joe and Tan, Xianying},
-	year = {2024},
-	date = {2024},
-	url = {https://github.com/rstudio/DT}
+@Manual{DT,
+  title = {DT: A Wrapper of the JavaScript Library 'DataTables'},
+  author = {Yihui Xie and Joe Cheng and Xianying Tan},
+  year = {2024},
+  note = {R package version 0.33},
+  url = {https://CRAN.R-project.org/package=DT},
+  doi = {10.32614/CRAN.package.DT},
 }
 
-@article{ggplot2,
-	title = {ggplot2: Elegant Graphics for Data Analysis},
-	author = {Wickham, Hadley},
-	year = {2016},
-	date = {2016},
-	url = {https://ggplot2.tidyverse.org}
+@Book{ggplot2,
+  title = {ggplot2: Elegant Graphics for Data Analysis},
+  author = {Hadley Wickham},
+  publisher = {Springer-Verlag New York},
+  year = {2016},
+  isbn = {978-3-319-24277-4},
+  url = {https://ggplot2.tidyverse.org},
+  doi = {10.1007/978-3-319-24277-4},
 }
 
-@article{plotly,
-	title = {Interactive Web-Based Data Visualization with R, plotly, and shiny},
-	author = {Sievert, Carson},
-	year = {2020},
-	date = {2020},
-	url = {https://plotly-r.com}
+@Book{plotly,
[…]
