### 2025-04-16 #8055 @DatViseR
The new build of the paper contains all the missing doi's and a new version of the bib file.  Sorry for this obvious mistake in the previous version.  All the best!
Tomasz Stępkowski

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1093/bib/bbw114 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: shiny: Web Application Framework for R
> - No DOI given, and none found for title: ggplot2: Elegant Graphics for Data Analysis
> - No DOI given, and none found for title: gt: Easily Create Presentation-Ready Display Table...
> - No DOI given, and none found for title: dplyr: A Grammar of Data Manipulation
>
> ❌ MISSING DOIs
>
> - 10.32614/cran.package.shiny.semantic may be a valid DOI for title: shiny.semantic: Semantic UI Support for Shiny
> - 10.32614/cran.package.dt may be a valid DOI for title: DT: A Wrapper of the JavaScript Library ’DataTable...
> - 10.1201/9780429447273 may be a valid DOI for title: Interactive Web-Based Data Visualization with R, p...
> - 10.32614/cran.package.shinyjs may be a valid DOI for title: shinyjs: Easily Improve the User Experience of You...
> - 10.32614/cran.package.shinyalert may be a valid DOI for title: shinyalert: Easily Create Pretty Popup Messages (M...
> - 10.32614/cran.package.tidyr may be a valid DOI for title: tidyr: Tidy Messy Data
>
> ❌ INVALID DOIs
>
> - None
> ```
