### 2025-04-17 a6deed0
inconsistency in capital letter in file vivid_functions.R repaired
