### 2025-06-09 9a2d7ba
Draft unix installation script introduced - additional info on system dependencies will be added
