+++ install_vivid_volcano.sh
@@ -1,0 +1,841 @@
+#!/bin/bash
+
+# ====================================================================
+# Vivid Volcano - Core Installation Script (PostgreSQL-free)
+# ====================================================================
+# Author: DatViseR
+# Date: 2025-06-09
+# Description: Clones Vivid Volcano repository and sets up core renv environment
+# Repository: https://github.com/DatViseR/Vivid-Volcano
+#  ====================================================================
+
+set -e  # Exit on any error
+
+# Colors for output
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+YELLOW='\033[1;33m'
+BLUE='\033[0;34m'
+PURPLE='\033[0;35m'
+CYAN='\033[0;36m'
+NC='\033[0m'
+
+# Logging functions
+print_header() {
+    echo -e "\n${PURPLE}=================================${NC}"
+    echo -e "${PURPLE}$1${NC}"
+    echo -e "${PURPLE}=================================${NC}\n"
+}
+
+print_step() {
+    echo -e "${BLUE}[STEP]${NC} $1"
+}
+
+print_info() {
+    echo -e "${CYAN}[INFO]${NC} $1"
+}
+
+print_success() {
+    echo -e "${GREEN}[SUCCESS]${NC} $1"
+}
+
+print_warning() {
+    echo -e "${YELLOW}[WARNING]${NC} $1"
+}
+
+print_error() {
+    echo -e "${RED}[ERROR]${NC} $1"
+}
+
+# Utility functions
+command_exists() {
+    command -v "$1" >/dev/null 2>&1
+}
+
+check_r_version() {
+    local r_version=$(R --version | head -n1 | grep -o 'R version [0-9]\+\.[0-9]\+\.[0-9]\+' | grep -o '[0-9]\+\.[0-9]\+\.[0-9]\+')
+    local major=$(echo $r_version | cut -d. -f1)
+    local minor=$(echo $r_version | cut -d. -f2)
+    
+    if [[ $major -gt 4 ]] || [[ $major -eq 4 && $minor -ge 4 ]]; then
+        return 0
+    else
+        return 1
+    fi
+}
+
+detect_os() {
+    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        if command_exists lsb_release; then
+            echo $(lsb_release -si | tr '[:upper:]' '[:lower:]')
+        elif [[ -f /etc/os-release ]]; then
+            echo $(grep '^ID=' /etc/os-release | cut -d= -f2 | tr -d '"' | tr '[:upper:]' '[:lower:]')
+        else
+            echo "linux"
+        fi
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        echo "macos"
+    elif [[ "$OSTYPE" == "cygwin" ]] || [[ "$OSTYPE" == "msys" ]]; then
+        echo "windows"
+    else
+        echo "unknown"
+    fi
+}
+
+# Essential system dependencies (minimal set)
+install_essential_dependencies() {
+    print_header "INSTALLING ESSENTIAL SYSTEM DEPENDENCIES"
+    
+    local os=$(detect_os)
+    print_info "Detected OS: $os"
+    
+    case $os in
+        ubuntu|debian)
+            print_step "Installing minimal Ubuntu/Debian dependencies"
+            if command_exists apt-get; then
+                print_info "Updating package list..."
+                sudo apt-get update -qq
+                
+                # Only essential packages for core R packages
+                local packages=(
+                    "libssl-dev"             # SSL library (for openssl, httr)
+                    "libcurl4-openssl-dev"   # cURL library (for curl, httr)
+                    "libxml2-dev"            # XML library (for xml2) - optional
+                    "libfontconfig1-dev"     # Font configuration (for Cairo) - optional
[…]
