### 2025-06-09 afb01cc
introduced description for the sustem dependencies to be installed optionally
