+++ install_vivid_volcano.sh
@@ -91,85 +91,184 @@ install_essential_dependencies() {
     
     case $os in
         ubuntu|debian)
-            print_step "Installing minimal Ubuntu/Debian dependencies"
+            print_step "Available Ubuntu/Debian dependencies"
             if command_exists apt-get; then
                 print_info "Updating package list..."
                 sudo apt-get update -qq
                 
-                # Only essential packages for core R packages
-                local packages=(
-                    "libssl-dev"             # SSL library (for openssl, httr)
-                    "libcurl4-openssl-dev"   # cURL library (for curl, httr)
-                    "libxml2-dev"            # XML library (for xml2) - optional
-                    "libfontconfig1-dev"     # Font configuration (for Cairo) - optional
-                    "libcairo2-dev"          # Cairo graphics (for Cairo) - optional
-                    "libharfbuzz-dev"        # Text shaping - optional
-                    "libfribidi-dev"         # Bidirectional text - optional
-                    "libfreetype6-dev"       # Font rendering - optional
-                    "libpng-dev"             # PNG support - optional
-                    "libjpeg-dev"            # JPEG support - optional
+                # Define packages with descriptions
+                declare -A packages_desc=(
+                    ["libssl-dev"]="SSL library - Required for secure connections (openssl, httr packages)"
+                    ["libcurl4-openssl-dev"]="cURL library - Required for HTTP requests (curl, httr packages)"
+                    ["libxml2-dev"]="XML library - Optional for XML processing (xml2 package)"
+                    ["libfontconfig1-dev"]="Font configuration - Optional for advanced text rendering"
+                    ["libcairo2-dev"]="Cairo graphics - Optional for high-quality graphics output"
+                    ["libharfbuzz-dev"]="Text shaping - Optional for complex text layout"
+                    ["libfribidi-dev"]="Bidirectional text - Optional for right-to-left text support"
+                    ["libfreetype6-dev"]="Font rendering - Optional for custom font support"
+                    ["libpng-dev"]="PNG support - Optional for PNG image processing"
+                    ["libjpeg-dev"]="JPEG support - Optional for JPEG image processing"
                 )
                 
-                print_info "Installing essential packages: ${packages[*]}"
-                sudo apt-get install -y "${packages[@]}" || {
-                    print_warning "Some optional packages failed, but continuing with core installation..."
-                }
+                local packages_to_install=()
                 
-                print_success "Essential dependencies installation completed"
+                echo
+                print_info "Please choose which dependencies to install:"
+                echo
+                
+                for package in "${!packages_desc[@]}"; do
+                    echo -e "${CYAN}Package:${NC} ${package}"
+                    echo -e "${YELLOW}Description:${NC} ${packages_desc[$package]}"
+                    read -p "Install $package? (Y/n): " -n 1 -r
+                    echo
+                    if [[ ! $REPLY =~ ^[Nn]$ ]]; then
+                        packages_to_install+=("$package")
+                        echo -e "${GREEN}✓ Will install $package${NC}"
+                    else
+                        echo -e "${YELLOW}⚬ Skipping $package${NC}"
+                    fi
+                    echo
+                done
+                
+                if [[ ${#packages_to_install[@]} -gt 0 ]]; then
+                    print_info "Installing selected packages: ${packages_to_install[*]}"
+                    sudo apt-get install -y "${packages_to_install[@]}" || {
+                        print_warning "Some packages failed to install, but continuing with core installation..."
+                    }
+                    print_success "Selected dependencies installation completed"
+                else
+                    print_info "No packages selected for installation"
+                fi
             else
[…]
