### 2025-06-10 b4429c4
Updated the readme with the thorough information about how to install Vivid Volcano and restore the environment ; create a contributing.md file
