+++ README.md
@@ -32,33 +32,170 @@ of detected and regulated genes within a specific category\
 • Process log system for user sessions, useful for checking data
 processing outcomes, debugging, and analyzing app functionality
 
-**The application is available in the cloud:**
+## **The application is available in the cloud or can be installed locally:**
 
-[**Vivid volcano live at POSIT Connect
-Cloud**](https://datviser-vivid-volcano.share.connect.posit.cloud//)
+### **Vivid Volcano live at POSIT Connect Cloud**
 
-**The live demo of the first beta version linked below. Now the app is
-after closed beta testing, pre-released and waiting for professional
+**Data Privacy Notice**\
+When you use this cloud-based application, your data is processed
+securely within your own session. What it means :
+
+-   **Your uploaded data is not stored permanently**, and it is **not
+    accessible to other users**.
+
+-   Each session is isolated, meaning your data is **only available
+    during your active session** and is **automatically cleared when the
+    session ends (including temporal logs)**.
+
+-   Vivid Volcano does **not collect, store, or share** any uploaded
+    files or analysis results. It collects non-sensitive telemetry
+    data - number of sessions, analyses performed, time of sessions
+    etc....
+
+-   For your safety, I recommend not uploading highly sensitive data
+    (ex. not blinded patients data) , as the app is hosted on a public
+    server without authentication.
+
+**The live demo of the first beta version is linked below. Now the app
+is after closed beta testing, pre-released and waiting for professional
 peer review from the scientific community.**
 
 [**Vivid volcano features- demo
 movie**](https://drive.google.com/file/d/1b7IzoJnTdTEW1VTk-L8TeDV9CMX0qKEz/view?usp=drive_link)
 
-To run the app locally follow the Instructions:
+### Instructions for running the app locally:
 
-1.  Clone this repository:\
-    git clone <https://github.com/DatViseR/Vivid-Volcano.git>
+Vivid Volcano uses **renv** (R environment management) to ensure
+reproducible package dependencies across different systems. This creates
+an isolated R environment with the exact package versions specified in
+`renv.lock`, preventing conflicts with your system-wide R packages.
 
-2.  Change directory:\
-    cd Vivid-Volcano
+#### For UNIX Systems (Linux/macOS)
 
-3.  Run the app in R:\
-    Rscript -e "shiny::runApp()"
+1.  Automated installation
 
-**Contributions are welcome!** Please fork the repository and submit a
-pull request.
+Download the [automated installation
+script](https://drive.google.com/file/d/11mdl1ZFYsPH2RoTNlqMMipiKJz9VmxAp/view?usp=sharing)
+written in bash - it that handles system dependencies, renv setup,
+cloning the repository and should enable working with Vivid Volcano
+locally with just few clicks.
 
-### How to cite Vivid Volcano?
+``` bash
+# Run the automated installer in the bash terminal
+chmod +x install_vivid_volcano.sh
+./install_vivid_volcano.sh
+```
+
+**What the installation script does:**
+
+-   Checks prerequisites (R 4.4+, Git, system dependencies)
+
+-   Offers selective installation of system libraries (SSL, cURL, Cairo,
+    etc.)
+
+-   Sets up renv environment with core packages
+
+-   Excludes problematic packages (PostgreSQL, V8) for stability
+
+-   Creates launcher scripts for easy app startup
+
+-   Generates an installation report
+
+2.  Manual installation
+
+If the installation script fails or you prefer full manual control over
[…]
