### 2025-06-21 21c244b
New version of installation script enabling direct start of app, louncher scripts erased
