+++ install_vivid_volcano.sh
@@ -4,8 +4,8 @@
 # Vivid Volcano - Core Installation Script (PostgreSQL-free)
 # ====================================================================
 # Author: DatViseR
-# Date: 2025-06-09
-# Description: Clones Vivid Volcano repository and sets up core renv environment
+# Date: 2025-06-21
+# Description: Clones Vivid Volcano repository, sets up core renv environment and runs the application
 # Repository: https://github.com/DatViseR/Vivid-Volcano
 #  ====================================================================
 
@@ -631,19 +631,16 @@ generate_summary <- function(install_success, verify_success, test_success) {
         }
     }
     
-    cat("\n=== NEXT STEPS ===\n")
+    cat("\n=== APPLICATION READY ===\n")
     if (status == "READY") {
-        cat("🎉 Core installation successful! You can now:\n")
-        cat("1. Run: ./run_vivid_volcano.sh\n")
-        cat("2. Or: R -e \"source('renv/activate.R'); shiny::runApp('app.R')\"\n")
+        cat("🎉 Core installation successful!\n")
+        cat("The application is ready to run in your browser.\n")
     } else if (status == "PARTIAL") {
-        cat("⚠ Partial installation. Try running the app - it may still work:\n")
-        cat("1. Run: ./run_vivid_volcano.sh\n")
-        cat("2. Some features may be limited\n")
+        cat("⚠ Partial installation. The app may still work.\n")
+        cat("Some features may be limited.\n")
     } else {
         cat("❌ Installation incomplete. Check error messages above.\n")
-        cat("1. Try running the script again\n")
-        cat("2. Or install packages manually in R\n")
+        cat("You may try running the app anyway.\n")
     }
     
     cat("\nNote: This core installation excludes PostgreSQL and other\n")
@@ -722,98 +719,6 @@ EOF
     fi
 }
 
-# Create simple launcher script
-create_simple_launcher() {
-    print_header "CREATING SIMPLE LAUNCHER SCRIPTS"
-    
-    print_step "Creating core launcher script"
-    
-    cat > run_vivid_volcano.sh << 'EOF'
-#!/bin/bash
-
-# Vivid Volcano Core Launcher
-# Focuses on core functionality without problematic packages
-
-echo "🌋 Starting Vivid Volcano (Core Version)..."
-
-# Check if we're in the right directory
-if [[ ! -f "app.R" ]] || [[ ! -f "renv.lock" ]]; then
-    echo "❌ Error: Not in Vivid Volcano directory"
-    echo "Please run this script from the Vivid-Volcano directory"
-    exit 1
-fi
-
-echo "Checking core environment..."
-
-R --slave --no-restore --no-save -e "
-# Activate renv
-source('renv/activate.R')
-
-# Check essential packages only
-essential <- c('shiny', 'dplyr', 'ggplot2', 'DT')
-missing <- c()
-
-for (pkg in essential) {
-    if (!requireNamespace(pkg, quietly = TRUE)) {
-        missing <- c(missing, pkg)
-    }
-}
-
-if (length(missing) > 0) {
-    cat('❌ Missing essential packages:', paste(missing, collapse = ', '), '\n')
-    cat('Please run the installation script again\n')
-    quit(status = 1)
-}
-
-cat('✅ Core environment verified\n')
-cat('🚀 Launching Vivid Volcano (Core)...\n')
-cat('📱 The app will open in your default web browser\n')
-cat('🛑 To stop the app, press Ctrl+C in this terminal\n\n')
-
-cat('Note: This core version focuses on volcano plots and GO analysis\n')
[…]
