### 2025-06-21 e52e5a9
Added link to the new installation script
