### 2025-06-30 1accf54
updated readme to include the instruction on running the app subseqent to first run after manual instalation, created a new renv lockfile from a newly set up environment from a fresh installation on an alternative system
