### 2025-06-30 #8055 @DatViseR
Dear @shrestha-c

Thank you for additional comments

> Thank you for the update. I have tested the automated installation on a Unix system, and it now works. Since many users may refer to the README first, it would be helpful to include the terminal command "./run_vivid_volcano.sh" there as well for clarity.

I added the README instruction on how to run the app after first automatic run after automatic installation:

"
Next time if you want to run the app from the unix terminal type:
``` bash
# Acivate R terminal and run Vivid-Volcano
R -e "shiny::runApp('app.R')"
```
or run it from R GUI, R Studio or other R IDE
"

> For the manual installation, I still face the same issue. When I run R -e "shiny::runApp()" on the terminal, the output stops at: "Listening on http://127.0.0.1:7964" The app does not open in the browser or proceed further.
>
I tested the installation on the clean environment in Windows and Linux and it works - maybe this is some sort of mac problem - i will test this on Mac as well.

> Finally, when trying to launch via RStudio, I encountered an issue with the renv environment—it appears that the project is missing a lockfile. Here’s the message I get:
>
> renv::activate() Loading required package: GOSemSim Restarting R session...
>
>     * Project '~/' loaded. [renv 1.1.4]
>
>     * The project is out-of-sync -- use `renv::status()` for details.
>
>
> > renv::restore()
> > Error: This project does not contain a lockfile.
> > Have you called `snapshot()` yet?
> > Traceback (most recent calls last):
> > 5: renv::restore()
> > 4: lockfile %||% renv_lockfile_load(project = project, strict = TRUE)
> > 3: renv_lockfile_load(project = project, strict = TRUE)
> > 2: abort(c("This project does not contain a lockfile.", i = "Have you called `snapshot()` yet?"))
> > 1: stop(fallback)
> > renv::status()
> > This project does not contain a lockfile.
> > Use `renv::snapshot()` to create a lockfile.
>
> Hope this helps! Thanks, Chandani

I do not understand this issue -  the project had the lockfile (https://github.com/DatViseR/Vivid-Volcano/commits/master/renv.lock) . To be sure I have now additionaly updated this lockfile to include the recent version of R which works fine in one of the Windows testing systems.

Could you please check wheter your project in R studio has the working directory set to Vivid-Volcano and that you call renv:restore() from this working directory and not from the renv folder (renv.lock is located inside the main folder).  Alternatively  could you please check wheter you do not have any nested renv projects activated etc...

Thank you for your comments. It seems there are no problems with installation on Linux and Windows now -  I will further test the manual and automatic installation on Mac and let you know if i came across any issues.
