### 2025-07-02 8f9f496
Add enhanced browser launching for macOS with fallback mechanisms
Co-authored-by: DatViseR <64078730+DatViseR@users.noreply.github.com>
