+++ install_vivid_volcano.sh
@@ -759,7 +759,10 @@ DATA FILES:
 - Demo Data: $(test -f www/demo_data.csv && echo "Present" || echo "Missing")
 
 TO RUN VIVID VOLCANO:
-Run: R -e "shiny::runApp()"
+Enhanced launcher (recommended): ./launch_app.sh
+R launcher with browser support: R -f launch_app.R
+Basic command: R -e "shiny::runApp()"
+From RStudio: Open and run app.R
 
 LIMITATIONS IN THE CORE MASTER VERSION designed to work locally:
 - No PostgreSQL database connectivity (Install dependencies manually and clone deployed with telemetry branch of the repository)
@@ -782,38 +785,124 @@ launch_application() {
         return 1
     fi
     
+    # Detect operating system for enhanced browser launching
+    local os_type="unknown"
+    if [[ "$OSTYPE" == "darwin"* ]]; then
+        os_type="macos"
+        print_info "🍎 macOS detected - using enhanced browser launching"
+    elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        os_type="linux"
+        print_info "🐧 Linux detected"
+    elif [[ "$OSTYPE" == "msys" ]] || [[ "$OSTYPE" == "cygwin" ]]; then
+        os_type="windows"
+        print_info "🪟 Windows detected"
+    fi
+    
     print_info "🌋 Starting Vivid Volcano (Core Version)..."
     print_info "📱 The app will open in your default web browser"
     print_info "🛑 To stop the app, press Ctrl+C in this terminal"
     
-    # Launch the application
-    R --slave --no-restore --no-save -e "
-        # Activate renv
-        source('renv/activate.R')
+    # Use enhanced launcher if available, otherwise fallback to basic launcher
+    if [[ -f "launch_app.R" ]]; then
+        print_info "🚀 Using enhanced launcher with improved browser support..."
         
-        # Check essential packages
-        essential <- c('shiny', 'dplyr', 'ggplot2', 'DT')
-        missing <- c()
+        R --slave --no-restore --no-save -f launch_app.R
+    else
+        print_info "🚀 Using basic launcher..."
         
-        for (pkg in essential) {
-            if (!requireNamespace(pkg, quietly = TRUE)) {
-                missing <- c(missing, pkg)
+        # Launch the application with enhanced browser launching for macOS
+        R --slave --no-restore --no-save -e "
+            # Activate renv
+            source('renv/activate.R')
+            
+            # Check essential packages
+            essential <- c('shiny', 'dplyr', 'ggplot2', 'DT')
+            missing <- c()
+            
+            for (pkg in essential) {
+                if (!requireNamespace(pkg, quietly = TRUE)) {
+                    missing <- c(missing, pkg)
+                }
             }
-        }
-        
-        if (length(missing) > 0) {
-            cat('❌ Missing essential packages:', paste(missing, collapse = ', '), '\n')
-            cat('Installation may be incomplete\n')
-            quit(status = 1)
-        }
-        
-        cat('✅ Core environment verified\n')
-        cat('🚀 Launching Vivid Volcano...\n\n')
-        
-        # Launch the application
-        library(shiny)
-        runApp('app.R', launch.browser = TRUE, host = '127.0.0.1')
-    "
+            
+            if (length(missing) > 0) {
+                cat('❌ Missing essential packages:', paste(missing, collapse = ', '), '\n')
+                cat('Installation may be incomplete\n')
+                quit(status = 1)
+            }
+            
+            cat('✅ Core environment verified\n')
+            cat('🚀 Launching Vivid Volcano...\n\n')
+            
+            # Load shiny library
+            library(shiny)
+            
+            # Enhanced browser launching for macOS
[…]
