### 2025-07-03 #8055 @Nikhila123456
Hi All,

Could you please anyone point me where I need to run this command ** @editorialbot generate my checklist** for generating the respective checklist for me?

Thanks,
Nikhila
