### 2025-07-04 c6e61b1
Update install_vivid_volcano.sh
Corrected command for internet test conection
