### 2025-07-04 #8055 @DatViseR
> Hi All,
>
> I am running into the error below while trying to install automated installation script. Could someone help me out?
>
> Error in library(Cairo) : there is no package called ‘Cairo’ Calls: runApp ... sourceUTF8 -> eval -> eval -> ..stacktraceon.. -> library Execution halted

Dear @Nikhila123456

Could you please make sure that during the instalation you enable installation of the following system dependency and check wheter you still have this error.
Ubuntu/Debian:

`    libcairo2-dev - "Cairo graphics - Optional for high-quality graphics output"`

Red Hat/Fedora:

`    cairo-devel - "Cairo graphics - Optional for high-quality graphics output"`

macOS:

`    cairo - "Cairo graphics - Optional for high-quality graphics output"`

I just tested the installation script one more time (this time in fully clean virtual github codespace (Linux container)) and everything worked fine. Best!
