### 2025-07-07 8f57176
Universal version of installation script including special feuters for container installation
