+++ install_vivid_volcano.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 
 # ====================================================================
-# Vivid Volcano - Core Installation Script (PostgreSQL-free)
+# Vivid Volcano - Universal Installation Script 
 # ====================================================================
 # Author: DatViseR
-# Date: 2025-06-21
-# Description: Clones Vivid Volcano repository, sets up core renv environment and runs the application
+# Date: 2025-07-07
+# Description: Universal installation for local and cloud environments
 # Repository: https://github.com/DatViseR/Vivid-Volcano
-#  ====================================================================
+# Works on: macOS, Linux, Codespaces, Gitpod, and other containers
+# ====================================================================
 
 set -e  # Exit on any error
 
@@ -64,8 +65,18 @@ check_r_version() {
     fi
 }
 
-detect_os() {
-    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+detect_environment() {
+    # Cloud environments
+    if [[ -n "${CODESPACE_NAME:-}" ]] || [[ -n "${GITHUB_CODESPACES:-}" ]]; then
+        echo "codespaces"
+    elif [[ -n "${GITPOD_WORKSPACE_ID:-}" ]]; then
+        echo "gitpod"
+    elif [[ -n "${COLAB_GPU:-}" ]]; then
+        echo "colab"
+    elif [[ -n "${JUPYTER_SERVER_ROOT:-}" ]]; then
+        echo "jupyter"
+    # Local environments
+    elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
         if command_exists lsb_release; then
             echo $(lsb_release -si | tr '[:upper:]' '[:lower:]')
         elif [[ -f /etc/os-release ]]; then
@@ -82,208 +93,314 @@ detect_os() {
     fi
 }
 
-# Essential system dependencies (minimal set)
-install_essential_dependencies() {
-    print_header "INSTALLING ESSENTIAL SYSTEM DEPENDENCIES"
-    
-    local os=$(detect_os)
-    print_info "Detected OS: $os"
-    
-    case $os in
-        ubuntu|debian)
-            print_step "Available Ubuntu/Debian dependencies"
-            if command_exists apt-get; then
-                print_info "Updating package list..."
-                sudo apt-get update -qq
-                
-                # Define packages with descriptions
-                declare -A packages_desc=(
-                    ["libssl-dev"]="SSL library - Required for secure connections (openssl, httr packages)"
-                    ["libcurl4-openssl-dev"]="cURL library - Required for HTTP requests (curl, httr packages)"
-                    ["libxml2-dev"]="XML library - Optional for XML processing (xml2 package)"
-                    ["libfontconfig1-dev"]="Font configuration - Optional for advanced text rendering"
-                    ["libcairo2-dev"]="Cairo graphics - Optional for high-quality graphics output"
-                    ["libharfbuzz-dev"]="Text shaping - Optional for complex text layout"
-                    ["libfribidi-dev"]="Bidirectional text - Optional for right-to-left text support"
-                    ["libfreetype6-dev"]="Font rendering - Optional for custom font support"
-                    ["libpng-dev"]="PNG support - Optional for PNG image processing"
-                    ["libjpeg-dev"]="JPEG support - Optional for JPEG image processing"
-                )
-                
-                local packages_to_install=()
-                
-                echo
-                print_info "Please choose which dependencies to install:"
-                echo
-                
-                for package in "${!packages_desc[@]}"; do
-                    echo -e "${CYAN}Package:${NC} ${package}"
-                    echo -e "${YELLOW}Description:${NC} ${packages_desc[$package]}"
-                    read -p "Install $package? (Y/n): " -n 1 -r
[…]
