### 2025-07-07 b6db987
wrong bollean case for R part of the script corrected
