### 2025-07-07 a94275d
additional automatic review for terminal clutter and MacOS dep handling introduced
