+++ install_vivid_volcano.sh
@@ -4,48 +4,89 @@
 # Vivid Volcano - Universal Installation Script 
 # ====================================================================
 # Author: DatViseR
-# Date: 2025-07-07
-# Description: Universal installation for local and cloud environments
+# Date: 2025-01-08
+# Description: Production-ready universal installer with enhanced macOS dependency handling
 # Repository: https://github.com/DatViseR/Vivid-Volcano
 # Works on: macOS, Linux, Codespaces, Gitpod, and other containers
 # ====================================================================
 
 set -e  # Exit on any error
 
-# Colors for output
+# Colors and formatting
 RED='\033[0;31m'
 GREEN='\033[0;32m'
 YELLOW='\033[1;33m'
 BLUE='\033[0;34m'
 PURPLE='\033[0;35m'
 CYAN='\033[0;36m'
+BOLD='\033[1m'
 NC='\033[0m'
 
 # Logging functions
 print_header() {
-    echo -e "\n${PURPLE}=================================${NC}"
-    echo -e "${PURPLE}$1${NC}"
-    echo -e "${PURPLE}=================================${NC}\n"
+    echo -e "\n${PURPLE}${BOLD}━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━${NC}"
+    echo -e "${PURPLE}${BOLD} $1${NC}"
+    echo -e "${PURPLE}${BOLD}━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━${NC}\n"
 }
 
 print_step() {
-    echo -e "${BLUE}[STEP]${NC} $1"
+    echo -e "${BLUE}▶${NC} $1"
 }
 
 print_info() {
-    echo -e "${CYAN}[INFO]${NC} $1"
+    echo -e "${CYAN}ℹ${NC} $1"
 }
 
 print_success() {
-    echo -e "${GREEN}[SUCCESS]${NC} $1"
+    echo -e "${GREEN}✓${NC} $1"
 }
 
 print_warning() {
-    echo -e "${YELLOW}[WARNING]${NC} $1"
+    echo -e "${YELLOW}⚠${NC} $1"
 }
 
 print_error() {
-    echo -e "${RED}[ERROR]${NC} $1"
+    echo -e "${RED}✗${NC} $1"
+}
+
+# Enhanced progress indicator with spinner
+show_progress() {
+    local message="$1"
+    local duration="${2:-60}"
+    local pid=$!
+    
+    local spinner='⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏'
+    local delay=0.1
+    local i=0
+    
+    echo -ne "${CYAN}⏳${NC} $message "
+    
+    while kill -0 $pid 2>/dev/null; do
+        local char="${spinner:$i:1}"
+        echo -ne "\r${CYAN}⏳${NC} $message ${char}"
+        sleep $delay
+        i=$(( (i+1) % ${#spinner} ))
+    done
+    
+    wait $pid
+    local exit_code=$?
+    
+    if [ $exit_code -eq 0 ]; then
+        echo -e "\r${GREEN}✓${NC} $message - completed successfully"
+    else
+        echo -e "\r${RED}✗${NC} $message - failed"
+    fi
+    
+    return $exit_code
+}
+
+# Enhanced command execution with progress
+execute_with_progress() {
+    local command="$1"
+    local message="$2"
+    
+    eval "$command" >/dev/null 2>&1 &
[…]
