### 2025-07-07 e4fef64
Should include all the essential dependencies separated from the optional
