+++ install_vivid_volcano.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 
 # ====================================================================
-# Vivid Volcano - Universal Installation Script 
+# Vivid Volcano - Universal Installation Script
 # ====================================================================
 # Author: DatViseR
 # Date: 2025-01-08
-# Description: Production-ready universal installer with enhanced macOS dependency handling
+# Description: Production-ready universal installer with corrected package categorization,
+# provides information on essential and optional (enhanced functions) dependencies
 # Repository: https://github.com/DatViseR/Vivid-Volcano
 # Works on: macOS, Linux, Codespaces, Gitpod, and other containers
 # ====================================================================
@@ -49,10 +50,9 @@ print_error() {
     echo -e "${RED}✗${NC} $1"
 }
 
-# Enhanced progress indicator with spinner
+# Enhanced progress indicator
 show_progress() {
     local message="$1"
-    local duration="${2:-60}"
     local pid=$!
     
     local spinner='⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏'
@@ -135,7 +135,7 @@ is_container_environment() {
     esac
 }
 
-# Enhanced R version check with error handling
+# Enhanced R version check
 check_r_version() {
     if ! command_exists R; then
         return 1
@@ -152,7 +152,7 @@ check_r_version() {
     major=$(echo "$r_version" | cut -d. -f1)
     minor=$(echo "$r_version" | cut -d. -f2)
     
-    if [[ $major -gt 4 ]] || [[ $major -eq 4 && $minor -ge 4 ]]; then
+    if [[ $major -gt 4 ]] || [[ $major -eq 4 && $minor -ge 3 ]]; then
         return 0
     else
         return 1
@@ -292,7 +292,7 @@ check_and_install_prerequisites() {
     else
         print_success "R available"
         if ! check_r_version; then
-            print_warning "R version <4.4 detected (4.4+ recommended)"
+            print_warning "R version <4.3 detected (4.3+ recommended for all packages)"
         fi
     fi
     
@@ -305,7 +305,7 @@ check_and_install_prerequisites() {
         fi
         
         if ! command_exists brew; then
-            print_warning "Homebrew not found (recommended for dependencies)"
+            print_warning "Homebrew not found (recommended for optional dependencies)"
             print_info "Install from: https://brew.sh"
         fi
     fi
@@ -362,23 +362,24 @@ install_system_dependencies() {
             
             # Essential dependencies (no choice)
             local essential_packages=("libssl-dev" "libcurl4-openssl-dev")
-            print_info "Installing essential libraries (required)"
+            print_info "Installing essential libraries (required for basic functionality)"
             execute_with_progress "sudo apt-get update -qq && sudo apt-get install -y ${essential_packages[*]}" "Installing essential libraries"
             
             # Optional dependencies with clear explanations
             echo
-            print_step "Optional system libraries (choose which features you want)"
+            print_step "Optional system libraries enhance Vivid Volcano features"
+            print_info "The app works without these, but they enable advanced functionality"
             echo
             
             local optional_deps=(
-                "libxml2-dev:XML data processing support:Enables xml2 R package for XML/HTML parsing"
-                "libcairo2-dev:High-quality graphics rendering:Enables Cairo R package for publication-quality plots"
-                "libfontconfig1-dev:Advanced font support:Better text rendering in plots"
-                "libharfbuzz-dev:Complex text layout:Support for complex scripts and typography"
-                "libfribidi-dev:Bidirectional text:Right-to-left text support (Arabic, Hebrew)"
-                "libfreetype6-dev:Font rendering engine:Enhanced font display"
[…]
