### 2025-07-07 3ba2946
The app now handles missing optional cairo and webshot dependencies
