### 2025-07-07 b6bfb3d
Enhanced info on R packadge life-time installation
