### 2025-07-08 37e7e6e
Included some important info for the installation script users
