### 2025-07-08 5c3503b
added information on Webshot2 dependency - optional for disabled gt tables pdf export functionality
