### 2025-07-08 961874a
Formated terminal info on cloud version of the app
