### 2025-07-08 cdc0e6a
Formated terminal info on cloud version of the app
