### 2025-07-22 #8055 @DatViseR
> Hi [@DatViseR](https://github.com/DatViseR),
>
> Great work. I could run the installation which successfully launched the app in browser from macOS. I am trying to figure out what I am supposed to be do next. Could you please guide me? Do I have to upload any any CSV file to ensure the functionality of the tool?
>
> Thanks, Nikhila

Dear @Nikhila123456 in case you do not have your own pre-processed omics data you can use the implemented "download demo data"  csv upload button to upload a small demo csv.
