### 2025-07-22 #8055 @DatViseR
> Hi [@DatViseR](https://github.com/DatViseR),
>
> The header columns in the CSV file were not recognized, even though the “Header” option was activated in the tool. Could you please look into this and let me know what needs to be done to resolve it?
>
> Thanks

Dear @Nikhila123456  I do not understand why you have this problem -  it works on many systems and the cloud. Are you sure you used this button to download and then used the proper decimal and separator as noted in the download button itself (separator is tab,  decimal is comma - it is a TSV file)?  Or have you tried on your data?  If so, could you please send at least the first rows so I can reproduce the issue...or could you describe the problem in more detail ?

To avoid confusion I clearly described the separator and decimal in the download button below:
