### 2025-08-05 a756e5d
Added a paragraph on automatic installer in the paper.md
