### 2025-08-05 35795c6
Update README.md
Updated the citation of the release
