+++ README.md
@@ -316,9 +316,7 @@ Non-Bioinformaticians to Analyze Pre-Processed Omics Data. Zenodo.
 
 **If you want to cite the source code image from Zenodo:**
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15119703.svg)](https://doi.org/10.5281/zenodo.15119703)
+[![DOI]([https://zenodo.org/badge/DOI/10.5281/zenodo.15119703.svg)](https://doi.org/10.5281/zenodo.15119703](https://doi.org/10.5281/zenodo.15119702))
 
-Tomasz Stępkowski. (2025). DatViseR/Vivid-Volcano - an intuitive tool
-that helps experimental scientists with no bioinformatics background
-explore and analyze pre-processed omics data. Pre-release. (v0.1.2).
-Zenodo. <https://doi.org/10.5281/zenodo.15119703>
+Tomasz Stępkowski. (2025). DatViseR/Vivid-Volcano: Vivid Volcano v1.0 - an intuitive tool that helps experimental scientists with no bioinformatics background explore and analyze pre-processed omics data. (v1.0.0). Zenodo.
+<https://doi.org/10.5281/zenodo.16742973>
