### 2025-08-05 #8055 @DatViseR
Dear @lpantano

Thank you for your editorial work. In response to the recent checklist, I have :

1. Checked the ORCID and affiliation - it is correct

3. Create a new release v1.0  (only the doi link to the Zenodo repository in the Readme was updated later when it was ready after release creation)

5. Updated the ORCID link and affiliation in Zenodo to match the affiliation in the paper

7. The licence in Zenodo and GitHub repository is marked as AGPL 3.0 (the draft article compiles properly, but does not include the proper licence yet)

> @DatViseR can you check the branch that we need to use to generate the pdf?

The branch is called Master not Main (it is the default one). I think if there was a problem with GitHub action workflow, it might have been due to an obsolete GitHub access TOKEN, which I have now updated and it seems the paper is properly compiled:

Please let me know if anything more is needed from my site.

Regards
Tomasz Stępkowski

> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
>     * Double check authors and affiliations (including ORCIDs)
>
>     * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>
>     * Archive the release on Zenodo/figshare/etc and post the DOI here.
>
>     * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>
>     * Make sure that the license listed for the archive is the same as the software license.
>
>
