### 2025-08-12 #8055 @Nikhila123456
@DatViseR,
I noticed that a few of the figures seem to be a bit low in resolution. You might consider using higher resolution, especially for multi-panel figures, so all details are clear. I also observed that in the Results section there’s a subsection titled GO Category Details, but I couldn’t find a figure that directly shows what’s described there. Please feel free to handle them in whatever way you think works best. Thank you!
