### 2025-08-21 378fba9
Introduce panel highresolution figures, change the title of the Zenodo source code repository
