+++ README.md
@@ -318,5 +318,5 @@ Non-Bioinformaticians to Analyze Pre-Processed Omics Data. Zenodo.
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15119703.svg)](https://doi.org/10.5281/zenodo.15119702)
 
-Tomasz Stępkowski. (2025). DatViseR/Vivid-Volcano: Vivid Volcano v1.0 - an intuitive tool that helps experimental scientists with no bioinformatics background explore and analyze pre-processed omics data. (v1.0.0). Zenodo.
+Tomasz Stępkowski. (2025). Vivid Volcano: Empowering Non-Bioinformaticians to Analyze Pre-Processed Omics Data. Zenodo. 
 <https://doi.org/10.5281/zenodo.16742973>
