### 2025-08-21 b9bf44f
Erased the figure numbering as this is the part of the  github action for producing paper
