### 2025-09-09 #8055 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] please ensure the archive title exactly matches the paper title
- [ ] licensing: The Zenodo archive lists GPL, while it seems on your repo it's AGPL, these need to be consistent\
- [ ] references: make sure you are {enforcing} capitalization where necessary in your .bib file (e.g. in Chang et al. and Sievert)

Finally, a bit of copyediting:
- [ ] Summary: either add a comma and remove the "and" after "volcano plots" or remove the comma after "exploration"
- [ ] Statement of Need, first sentence: weird extra space after "etc." in first parenthetical, missing space before citation at end of sentence
- [ ] Statement of Need, second paragraph: either format as a numbered list (i.e. with separate lines for each entry) or instead of including numerals, rephrase to "First, empowering..." and then "Second, lowering..."
- [ ] same paragraph: missing period at end of sentence
- [ ] first full paragraph on second page: add comma after "In summary"
- [ ] last paragraph of Statement of Need: capitalize Windows
- [ ] Technical description section: fix spacing before citations and duplication of Wickham (also no spaces before punctuation)
