### 2025-09-10 0736520
Corrections to the manuscript based on last JOSS  editorial remarks
