+++ Publication/paper.bib
@@ -1,107 +1,107 @@
 
 @article{manzoni2018,
-	title = {Genome, transcriptome and proteome: the rise of omics data and their integration in biomedical sciences},
-	author = {Manzoni, Claudia and Kia, Demis A and Vandrovcova, Jana and Hardy, John and Wood, Nicholas W and Lewis, Patrick A and Ferrari, Raffaele},
-	year = {2018},
-	month = {03},
-	date = {2018-03-01},
-	journal = {Briefings in Bioinformatics},
-	pages = {286--302},
-	volume = {19},
-	number = {2},
-	doi = {10.1093/bib/bbw114},
-	url = {https://academic.oup.com/bib/article/19/2/286/2562648},
-	langid = {en}
+  title   = {Genome, transcriptome and proteome: the rise of omics data and their integration in biomedical sciences},
+  author  = {Manzoni, Claudia and Kia, Demis A and Vandrovcova, Jana and Hardy, John and Wood, Nicholas W and Lewis, Patrick A and Ferrari, Raffaele},
+  date    = {2018-03-01},
+  year    = {2018},
+  month   = {03},
+  journal = {Briefings in Bioinformatics},
+  volume  = {19},
+  number  = {2},
+  pages   = {286--302},
+  doi     = {10.1093/bib/bbw114},
+  url     = {https://academic.oup.com/bib/article/19/2/286/2562648},
+  langid  = {en}
 }
 
 @Manual{shiny,
-  title = {shiny: Web Application Framework for R},
+  title  = {{shiny}: Web Application Framework for {R}},
   author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
-  year = {2024},
-  note = {R package version 1.8.1},
-  url = {https://CRAN.R-project.org/package=shiny},
-  doi = {10.32614/CRAN.package.shiny},
+  year   = {2024},
+  note   = {R package version 1.8.1},
+  url    = {https://CRAN.R-project.org/package=shiny},
+  doi    = {10.32614/CRAN.package.shiny}
 }
 
 @Manual{shiny.semantic,
-  title = {shiny.semantic: Semantic UI Support for Shiny},
+  title  = {{shiny.semantic}: {Semantic UI} Support for {Shiny}},
   author = {Filip Stachura and Dominik Krzeminski and Krystian Igras and Adam Forys and {Paweł Przytuła} and Jakub Chojna and Olga Mierzwa-Sulima and Jakub Nowicki and Tymoteusz Makowski},
-  year = {2024},
-  note = {R package version 0.4.4},
-  url = {https://CRAN.R-project.org/package=shiny.semantic},
-  doi = {10.32614/CRAN.package.shiny.semantic},
+  year   = {2024},
+  note   = {R package version 0.4.4},
+  url    = {https://CRAN.R-project.org/package=shiny.semantic},
+  doi    = {10.32614/CRAN.package.shiny.semantic}
 }
 
 @Manual{DT,
-  title = {DT: A Wrapper of the JavaScript Library 'DataTables'},
+  title  = {{DT}: A Wrapper of the {JavaScript} Library {DataTables}},
   author = {Yihui Xie and Joe Cheng and Xianying Tan},
-  year = {2024},
-  note = {R package version 0.33},
-  url = {https://CRAN.R-project.org/package=DT},
-  doi = {10.32614/CRAN.package.DT},
+  year   = {2024},
+  note   = {R package version 0.33},
+  url    = {https://CRAN.R-project.org/package=DT},
[…]
