### 2025-09-10 #8055 @DatViseR
Dear @rkurchin

Thank you for the final editorial assessment - please find the descriptions of suggested corrections below - I hope all the issues raised are now solved.

**Editorial issues:**

> -   Archive listed title and authors matches paper -> please ensure the archive title exactly matches the paper title

*Corrected - inserted the middle name abbreviation "M." in the archive and erased "DatViseR" from the front of archive title*

> -   Archive listed license matches software license

*Corrected - GPL in the archive changed to AGPL and now matches the repository license*

> -   Checked reference rendering –\> make sure you are {enforcing} capitalisation where necessary in your .bib file (e.g. in Chang et al. and Sievert)

*Corrected - The reference file has been modified. The capitalisation and case of some words such as "R" vs "r" in Sievert et al. and others was corrected*

**Copyediting:**
> -    Summary: either add a comma and remove the "and" after "volcano plots" or remove the comma after "exploration"

 *Corrected- comma was erased*

> -   Statement of Need, first sentence: weird extra space after "etc." in first parenthetical, missing space before citation at end of sentence

*Corrected - extra space erased, space added before citation parenthesis*

> -   Statement of Need, second paragraph: either format as a numbered list (i.e. with separate lines for each entry) or instead of including numerals, rephrase to "First, empowering..." and then "Second, lowering..."

*Corrected with the second option*

> -   same paragraph: missing period at end of sentence

*Corrected*

> -   *first full paragraph on second page: add comma after "In summary"*

*Corrected*

> -   last paragraph of Statement of Need: capitalize Windows -

 *Corrected*

> -   Technical description section: fix spacing before citations and duplication of Wickham (also no spaces before punctuation)

*Corrected - the references to Wickham were separated accordingly to the text and spacing fixed.
