### 2025-09-11 e5946c2
Added CITATION.cff file for JOSS article on Vivid Volcano
