### 2025-09-11 d744952
Revise citation details in README.md
Updated citation information for Vivid Volcano application.
